+++ lua/github-theme/util.lua
@@ -146,7 +146,7 @@ util.terminal = function(colors)
   vim.g.terminal_color_8 = colors.fg_dark
   -- light
   vim.g.terminal_color_7 = colors.fg_dark
-  vim.g.terminal_color_15 = colors.term_fg
+  vim.g.terminal_color_15 = colors.fg_term
   -- colors
   vim.g.terminal_color_1 = colors.red
   vim.g.terminal_color_9 = colors.bright_red