+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - refactor: modules export & typing
 - refactor: terminal themes moved 'extras' -> 'terminal'
 - feat: Apply highlight's style using `types.HighlightStyle` type
+- refactor: color name changed 'term_fg' -> 'fg_term'
 
 ### Fixes